requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.12.0
django-filter==2.4.0
django-import-export==2.5.0
django-phonenumber-field==5.2.0
django-restframework==0.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
phonenumbers==8.12.26
Pillow==8.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
qrcode==7.2
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
uvicorn==0.12.3
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
